Revert "Platform meltdown animation"
This reverts commit bdf360f1f86a7c7662f5402a8bcb7663ec7ef420.
