--- Popcorn/Popcorn.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28729.10
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Popcorn", "Popcorn\Popcorn.vcxproj", "{C4958476-B624-411F-AD91-D6918C7A280E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Debug|x64.ActiveCfg = Debug|x64
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Debug|x64.Build.0 = Debug|x64
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Debug|x86.ActiveCfg = Debug|Win32
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Debug|x86.Build.0 = Debug|Win32
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Release|x64.ActiveCfg = Release|x64
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Release|x64.Build.0 = Release|x64
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Release|x86.ActiveCfg = Release|Win32
-		{C4958476-B624-411F-AD91-D6918C7A280E}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A6FFC0D8-5980-4765-B80A-4C1C7540FE96}
-	EndGlobalSection
-EndGlobal
