--- Popcorn/Popcorn/Border.cpp
@@ -1,64 +1,0 @@
-﻿#include "Border.h"
-
-// AsBorder
-//------------------------------------------------------------------------------------------------------------
-AsBorder::AsBorder()
-: Border_Blue_Pen(0), Border_White_Pen(0), Border_Blue_Brush(0), Border_White_Brush(0)
-{
-}
-//------------------------------------------------------------------------------------------------------------
-void AsBorder::Init()
-{
-	AsConfig::Create_Pen_Brush(85, 255, 255, Border_Blue_Pen, Border_Blue_Brush);
-	AsConfig::Create_Pen_Brush(255, 255, 255, Border_White_Pen, Border_White_Brush);
-}
-//------------------------------------------------------------------------------------------------------------
-void AsBorder::Draw(HDC hdc, RECT &paint_area)
-{// Рисует рамку уровня
-
-	int i;
-
-	// 1. Линия слева
-	for (i = 0; i < 50; i++)
-		Draw_Element(hdc, 2, 1 + i * 4, false);
-
-	// 2. Линия справа
-	for (i = 0; i < 50; i++)
-		Draw_Element(hdc, 201, 1 + i * 4, false);
-
-	// 3. Линия сверху
-	for (i = 0; i < 50; i++)
-		Draw_Element(hdc, 3 + i * 4, 0, true);
-}
-//------------------------------------------------------------------------------------------------------------
-void AsBorder::Draw_Element(HDC hdc, int x, int y, bool top_border)
-{// Рисует элемент рамки уровня
-
- // Основная линия
-	SelectObject(hdc, Border_Blue_Pen);
-	SelectObject(hdc, Border_Blue_Brush);
-
-	if (top_border)
-		Rectangle(hdc, x * AsConfig::Global_Scale, (y + 1) * AsConfig::Global_Scale, (x + 4) * AsConfig::Global_Scale, (y + 4) * AsConfig::Global_Scale);
-	else
-		Rectangle(hdc, (x + 1) * AsConfig::Global_Scale, y * AsConfig::Global_Scale, (x + 4) * AsConfig::Global_Scale, (y + 4) * AsConfig::Global_Scale);
-
-	// Белая кайма
-	SelectObject(hdc, Border_White_Pen);
-	SelectObject(hdc, Border_White_Brush);
-
-	if (top_border)
-		Rectangle(hdc, x * AsConfig::Global_Scale, y * AsConfig::Global_Scale, (x + 4) * AsConfig::Global_Scale, (y + 1) * AsConfig::Global_Scale);
-	else
-		Rectangle(hdc, x * AsConfig::Global_Scale, y * AsConfig::Global_Scale, (x + 1) * AsConfig::Global_Scale, (y + 4) * AsConfig::Global_Scale);
-
-	// Перфорация
-	SelectObject(hdc, AsConfig::BG_Pen);
-	SelectObject(hdc, AsConfig::BG_Brush);
-
-	if (top_border)
-		Rectangle(hdc, (x + 2) * AsConfig::Global_Scale, (y + 2) * AsConfig::Global_Scale, (x + 3) * AsConfig::Global_Scale, (y + 3) * AsConfig::Global_Scale);
-	else
-		Rectangle(hdc, (x + 2) * AsConfig::Global_Scale, (y + 1) * AsConfig::Global_Scale, (x + 3) * AsConfig::Global_Scale, (y + 2) * AsConfig::Global_Scale);
-}
-//------------------------------------------------------------------------------------------------------------
