--- Popcorn/Popcorn/Border.h
@@ -1,20 +1,0 @@
-﻿#pragma once
-
-#include "Config.h"
-
-//------------------------------------------------------------------------------------------------------------
-class AsBorder
-{
-public:
-	AsBorder();
-
-	void Init();
-	void Draw(HDC hdc, RECT &paint_area);
-
-private:
-	void Draw_Element(HDC hdc, int x, int y, bool top_border);
-
-	HPEN Border_Blue_Pen, Border_White_Pen;
-	HBRUSH Border_Blue_Brush, Border_White_Brush;
-};
-//------------------------------------------------------------------------------------------------------------
