--- Popcorn/Popcorn/Main.h
@@ -1,4 +1,0 @@
-﻿#pragma once
-
-#include "resource.h"
-#include "Engine.h"
